add TabBar widget menu
