+++ lib/skeleton.dart
@@ -1,61 +1,36 @@
-import 'package:fitness_stopwatch_app/pages/home_pages.dart';
+import 'package:fitness_stopwatch_app/pages/timer_page.dart';
 import 'package:fitness_stopwatch_app/pages/parameter_page.dart';
 import 'package:flutter/material.dart';
 
-class MyHomePage extends StatefulWidget {
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
+class MyHomePage extends StatelessWidget {
+  const MyHomePage({super.key});
 
-class _MyHomePageState extends State<MyHomePage> {
-  var selectedIndex = 0;
   @override
   Widget build(BuildContext context) {
-    Widget page;
-    switch (selectedIndex) {
-      case 0:
-        page = HomePage();
-        break;
-      case 1:
-        page = ParameterPage();
-        break;
-      default:
-        throw UnimplementedError('no widget for $selectedIndex');
-    }
-    return LayoutBuilder(builder: (context, constraints) {
-      return Scaffold(
-        body: Row(
-          children: [
-            SafeArea(
-              child: NavigationRail(
-                extended: constraints.maxWidth >= 600,
-                destinations: [
-                  NavigationRailDestination(
-                    icon: Icon(Icons.timer),
-                    label: Text('Home'),
-                  ),
-                  NavigationRailDestination(
-                    icon: Icon(Icons.settings),
-                    label: Text('Settings'),
-                  ),
-                ],
-                selectedIndex: selectedIndex,
-                onDestinationSelected: (value) {
-                  setState(() {
-                    selectedIndex = value;
-                  });
-                },
+    return DefaultTabController(
+      initialIndex: 0,
+      length: 2,
+      child: Scaffold(
+        appBar: AppBar(
+          title: const Text('Fitness App'),
+          bottom: const TabBar(
+            tabs: <Widget>[
+              Tab(
+                icon: Icon(Icons.timer),
               ),
-            ),
-            Expanded(
-              child: Container(
-                color: Theme.of(context).colorScheme.primaryContainer,
-                child: page,
+              Tab(
+                icon: Icon(Icons.settings),
               ),
-            ),
+            ],
+          ),
+        ),
+        body: const TabBarView(
+          children: <Widget>[
+            TimerPage(),
+            ParameterPage(),
           ],
         ),
-      );
-    });
+      ),
+    );
   }
 }
